ModificacionesDelCliente V1.0

+++ CapaDatos/Dominio/Configuracion.cs
@@ -29,7 +29,7 @@ namespace CapaDatos.Dominio
         public string Facebook { get; set; }
         public int DescuentoSocios { get; set; }
         public int DescuentoPresencial { get; set; }
-        public int PorcentajeDescuentoHoy { get; set; }
+        public decimal PorcentajeDescuentoHoy { get; set; }
 
     }
 }
